Fix all clang warnings: unused captures, int conversions, shadow, double-promotion
Resolve 11 warnings surfaced by building with clang and -Wall -Wextra
-Wpedantic -Wconversion -Wshadow -Wdouble-promotion -Wsign-conversion:

- Remove unused 'this' lambda capture in session_analyzer.cpp
- Wrap bitwise-OR results in static_cast<uint16_t>() to silence
  -Wimplicit-int-conversion in test_generated_protocols.cpp
- Fix bgen codegen shadow pragma to cover clang (clang defines __GNUC__
  but was excluded by !__clang__ guard) in cpp_structs_decode.cpp
- Use explicit static_cast<double>() for float args to Catch2 WithinAbs
  to silence -Wdouble-promotion in test_bit_reader/writer.cpp
- Remove unused 'consumed' variable in test_bounded_queue.cpp
- Remove unused 'peer' variable in test_handler_registry.cpp

Verified: zero warnings on both clang 18 and GCC 13, all 1465 tests pass.

+++ conduit/tests/test_bit_reader.cpp
@@ -195,7 +195,7 @@ TEST_CASE("BitReader float reads", "[bit_reader]") {
         BitReader reader(data);
         auto val = reader.read_f32(Endian::Big);
         REQUIRE(val.has_value());
-        CHECK_THAT(*val, Catch::Matchers::WithinAbs(3.14f, 0.001));
+        CHECK_THAT(static_cast<double>(*val), Catch::Matchers::WithinAbs(static_cast<double>(3.14f), 0.001));
     }
 
     SECTION("read_f64") {
@@ -229,7 +229,7 @@ TEST_CASE("BitReader float reads little-endian", "[bit_reader]") {
         BitReader reader(data);
         auto val = reader.read_f32(Endian::Little);
         REQUIRE(val.has_value());
-        CHECK_THAT(*val, Catch::Matchers::WithinAbs(3.14f, 0.001));
+        CHECK_THAT(static_cast<double>(*val), Catch::Matchers::WithinAbs(static_cast<double>(3.14f), 0.001));
     }
 
     SECTION("read_f64 little-endian") {
@@ -259,7 +259,7 @@ TEST_CASE("BitReader float reads little-endian", "[bit_reader]") {
         BitReader reader(data);
         auto val = reader.read_f32(Endian::Little);
         REQUIRE(val.has_value());
-        CHECK_THAT(*val, Catch::Matchers::WithinAbs(-42.5f, 0.001));
+        CHECK_THAT(static_cast<double>(*val), Catch::Matchers::WithinAbs(static_cast<double>(-42.5f), 0.001));
     }
 
     SECTION("read_f64 little-endian roundtrip via writer") {
